@@ -1,62 +1,65 @@
 {
- "name": "uxcore-select-form-field",
- "version": "0.3.12",
- "description": "uxcore-select-form-field component for uxcore.",
- "repository": "https://github.com/uxcore/uxcore-select-form-field.git",
- "author": "eternalsky",
- "main": "build/index.js",
- "scripts": {
-  "start": "uxcore-tools run start",
-  "server": "uxcore-tools run server",
-  "lint": "uxcore-tools run lint",
-  "build": "uxcore-tools run build",
-  "test": "uxcore-tools run test",
-  "coverage": "uxcore-tools run coverage",
-  "pub": "uxcore-tools run pub",
-  "dep": "uxcore-tools run dep",
-  "tnpm-dep": "uxcore-tools run tnpm-dep",
-  "chrome": "uxcore-tools run chrome",
-  "browsers": "uxcore-tools run browsers",
-  "saucelabs": "uxcore-tools run saucelabs",
-  "update": "uxcore-tools run update"
- },
- "bugs": {
-  "url": "http://github.com/uxcore/uxcore-select-form-field/issues"
- },
- "keywords": [
-  "react",
-  "react-component",
-  "uxcore-select-form-field",
-  "SelectFormField",
-  ""
- ],
- "devDependencies": {
-  "console-polyfill": "^0.2.2",
-  "es5-shim": "^4.5.8",
-  "expect.js": "~0.3.1",
-  "kuma-base": "1.x",
-  "react": "0.14.x",
-  "react-addons-test-utils": "0.14.x",
-  "react-dom": "0.14.x",
-  "uxcore-button": "~0.4.3",
-  "uxcore-const": "0.0.2",
-  "uxcore-form": "~1.3.2",
-  "uxcore-kuma": "*",
-  "uxcore-tools": "0.2.x",
-  "jquery": "^3.1.1",
-  "enzyme": "^2.7.0"
- },
- "dependencies": {
-  "classnames": "^2.1.2",
-  "lie": "^3.1.0",
-  "lodash": "^4.15.0",
-  "natty-fetch": "^2.1.2",
-  "object-assign": "^4.0.0",
-  "uxcore-const": "0.0.2",
-  "uxcore-form-field": "~0.2.0",
-  "uxcore-select2": "~0.4.0",
-  "uxcore-validator": "~0.1.4"
- },
- "contributors": [],
- "license": "MIT"
+  "name": "uxcore-select-form-field",
+  "version": "0.3.12",
+  "description": "uxcore-select-form-field component for uxcore.",
+  "repository": "https://github.com/uxcore/uxcore-select-form-field.git",
+  "author": "eternalsky",
+  "main": "build/index.js",
+  "scripts": {
+    "start": "uxcore-tools run start",
+    "server": "uxcore-tools run server",
+    "lint": "uxcore-tools run lint",
+    "build": "uxcore-tools run build",
+    "test": "uxcore-tools run test",
+    "coverage": "uxcore-tools run coverage",
+    "pub": "uxcore-tools run pub",
+    "dep": "uxcore-tools run dep",
+    "tnpm-dep": "uxcore-tools run tnpm-dep",
+    "chrome": "uxcore-tools run chrome",
+    "browsers": "uxcore-tools run browsers",
+    "saucelabs": "uxcore-tools run saucelabs",
+    "update": "uxcore-tools run update",
+    "tnpm-update": "uxcore-tools run tnpm-update"
+  },
+  "bugs": {
+    "url": "http://github.com/uxcore/uxcore-select-form-field/issues"
+  },
+  "keywords": [
+    "react",
+    "react-component",
+    "uxcore-select-form-field",
+    "SelectFormField",
+    ""
+  ],
+  "devDependencies": {
+    "console-polyfill": "^0.2.2",
+    "enzyme": "^3.0.0",
+    "enzyme-adapter-react-15": "^1.0.0",
+    "es5-shim": "^4.5.8",
+    "expect.js": "~0.3.1",
+    "kuma-base": "1.x",
+    "react": "15.x",
+    "react-addons-test-utils": "15.x",
+    "react-dom": "15.x",
+    "react-test-renderer": "15.x",
+    "uxcore-button": "~0.4.3",
+    "uxcore-const": "0.0.2",
+    "uxcore-form": "~1.3.2",
+    "uxcore-kuma": "*",
+    "uxcore-tools": "0.2.x"
+  },
+  "dependencies": {
+    "classnames": "^2.1.2",
+    "lie": "^3.1.0",
+    "lodash": "^4.15.0",
+    "natty-fetch": "^2.1.2",
+    "object-assign": "^4.0.0",
+    "prop-types": "15.x",
+    "uxcore-const": "0.0.2",
+    "uxcore-form-field": "~0.2.0",
+    "uxcore-select2": "~0.4.0",
+    "uxcore-validator": "~0.1.4"
+  },
+  "contributors": [],
+  "license": "MIT"
 }
